backend/gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
PyJWT[crypto]
